feat: on_close_channel logic

+++ contracts/consumer/converter/src/contract.rs
@@ -75,7 +75,7 @@ impl ConverterContract<'_> {
         virtual_staking_code_id: u64,
         tombstoned_unbond_enable: bool,
         admin: Option<String>,
-        max_retrieve: u16,
+        max_retrieve: u32,
     ) -> Result<custom::Response, ContractError> {
         nonpayable(&ctx.info)?;
         // validate args
